+++ rio-backend/src/config/keyboard.rs
@@ -1,12 +1,9 @@
 use serde::{Deserialize, Serialize};
 
-use super::defaults::{default_bool_true, default_disable_ctlseqs_alt};
+use super::defaults::default_disable_ctlseqs_alt;
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone, Copy)]
 pub struct Keyboard {
-    // Enable kitty keyboard protocol
-    #[serde(default = "default_bool_true", rename = "use-kitty-keyboard-protocol")]
-    pub use_kitty_keyboard_protocol: bool,
     // Disable ctlseqs with ALT keys
     // For example: Terminal.app does not deal with ctlseqs with ALT keys
     #[serde(
@@ -20,7 +17,6 @@ pub struct Keyboard {
 impl Default for Keyboard {
     fn default() -> Keyboard {
         Keyboard {
-            use_kitty_keyboard_protocol: true,
             #[cfg(target_os = "macos")]
             disable_ctlseqs_alt: true,
             #[cfg(not(target_os = "macos"))]